refactor: trim dupe tests
